feat: update alchemy hooks
