+++ packages/alchemy/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../packages/tsconfig/node16.json",
-  "include": ["./src/**/*", "src/lib/utils.ts"],
+  "include": ["./src/**/*", "src/lib/utils.ts", "webhooks/activity.ts"],
   "exclude": ["node_modules"],
   "compilerOptions": {
     "sourceMap": true,
